+++ src/main/java/dto/Pet.java
@@ -1,3 +1,7 @@
+package dto;
+
+import dto.AbstractAnimal;
+
 public abstract class Pet extends AbstractAnimal {
 
 }
