+++ src/main/java/dto/Predator.java
@@ -1,3 +1,7 @@
+package dto;
+
+import dto.AbstractAnimal;
+
 public abstract class Predator extends AbstractAnimal {
 
 }
