+++ src/main/java/model/Dog.java
@@ -1,10 +1,14 @@
+package model;
+
+import dto.*;
+
 import java.math.BigDecimal;
 
-public class Shark extends Predator {
-    // голос акулы
-    private static final String voice = "Буль";
+public class Dog extends Pet {
+    // голос собаки
+    private static final String voice = "Гав";
 
-    public Shark(String breed, String name, BigDecimal cost, String character) {
+    public Dog(String breed, String name, BigDecimal cost, String character) {
         this.breed = breed;
         this.name = name;
         this.cost = cost;
