+++ src/main/java/model/Wolf.java
@@ -1,3 +1,7 @@
+package model;
+
+import dto.*;
+
 import java.math.BigDecimal;
 
 public class Wolf extends Predator {
